Jest: Move guilds into core
