--- src/util/guilds.ts
@@ -1,11 +1,0 @@
-import { UtilDao } from "./dao";
-import { Guild } from "discord.js";
-
-export async function getGuildId(guild: Guild): Promise<number> {
-	let botGuild = await UtilDao.getGuild(guild.id);
-	if (!botGuild) {
-		botGuild = await UtilDao.createGuild(guild.id, guild.name);
-	}
-
-	return botGuild.id;
-}
